[262] add test cases for HabitController - Renamed package
