+++ core/src/test/java/greencity/controller/HabitControllerTest.java
@@ -19,8 +19,8 @@ import greencity.dto.shoppinglistitem.ShoppingListItemDto;
 import greencity.dto.user.UserProfilePictureDto;
 import greencity.dto.user.UserVO;
 import greencity.exception.exceptions.BadRequestException;
-import greencity.reslver.PageableResolver;
-import greencity.reslver.UserVoResolver;
+import greencity.resolver.PageableResolver;
+import greencity.resolver.UserVoResolver;
 import greencity.service.HabitService;
 import greencity.service.LanguageService;
 import greencity.service.TagsService;
