+++ core/src/test/java/greencity/resolver/BaseArgumentResolver.java
@@ -1,4 +1,4 @@
-package greencity.reslver;
+package greencity.resolver;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
